refactor: select menu from ConsoleApp
